Windows: Tidy up daemon\utils*.go
