--- daemon/utils_linux.go
@@ -1,17 +1,0 @@
-// +build linux
-
-package daemon
-
-import "github.com/docker/libcontainer/selinux"
-
-func selinuxSetDisabled() {
-	selinux.SetDisabled()
-}
-
-func selinuxFreeLxcContexts(label string) {
-	selinux.FreeLxcContexts(label)
-}
-
-func selinuxEnabled() bool {
-	return selinux.SelinuxEnabled()
-}
