+++ daemon/utils_nounix.go
